--- main.js
@@ -1,6 +1,0 @@
-import "/style/reset.css";
-import "/style/global.css";
-
-import App from "./App.js";
-
-new App({ $target: document.querySelector("#app") });
